+++ app/components/source-editor.vue
@@ -40,5 +40,6 @@
 	{
 		resize: horizontal;
 		height: 100%;
+		background-color: #fffa;
 	}
 </style>